Module triple_accel::hamming

This module provides many Hamming distance routines.

These distance functions share the same efficient underlying SIMD-accelerated implementation:

hamming

hamming_simd_parallel

These search functions share the same efficient underlying SIMD-accelerated implementation:

hamming_search

hamming_search_simd

hamming_search_simd_with_opts

Functions

hamming
Returns the hamming distance between two strings using the best method.

hamming_naive
Returns the hamming distance between two strings by naively counting mismatches.

hamming_search
Returns an iterator over best Matchs by searching through the text haystack for the pattern needle using SIMD.

hamming_search_naive
Returns an iterator over best Matchs by naively searching through the text haystack for the pattern needle.

hamming_search_naive_with_opts
Returns an iterator over Matchs by naively searching through the text haystack for the pattern needle, with extra options.

hamming_search_simd
Returns an iterator over best Matchs by searching through the text haystack for the pattern needle using SIMD.

hamming_search_simd_with_opts
Returns an iterator over Matchs by searching through the text haystack for the pattern needle using SIMD, with extra options.

hamming_simd_movemask
Returns the hamming distance between two strings by counting mismatches using the SIMD movemask intrinsic.

hamming_simd_parallel
Returns the hamming distance between two strings by counting mismatches using SIMD vectors to increment multiple counters in parallel.

hamming_words_64
Returns the hamming distance between two strings by efficiently counting mismatches in chunks of 64 bits.

hamming_words_128
Returns the hamming distance between two strings by counting mismatches in chunks of 128 bits.