Function triple_accel::hamming::hamming_words_128

pub fn hamming_words_128(a: &[u8], b: &[u8]) -> u32

Returns the hamming distance between two strings by counting mismatches in chunks of 128 bits.

The length of a and b must be the same. Both a and b must be aligned and padded so they can be directly casted to chunks of u128. Use alloc_str to create aligned and padded strings. This may be slower than hamming_words_64 in practice, probably since Rust u128 is not as optimized. This should be slower than hamming_simd_parallel/movemask.

Arguments

a - first string (slice)

b - second string (slice)

Panics

If the length of a does not equal the length of b.

Example

let mut a = alloc_str(3);
let mut b = alloc_str(3);
fill_str(&mut a, b"abc");
fill_str(&mut b, b"abd");

let dist = hamming_words_128(&a, &b);

assert!(dist == 1);